Add missing variable CONTAINER_IMAGE_TO_TEST in Leap 15.5
If this variable is not set, it will try to fetch the container image
from lib/containers/urls.pm which is outdated.
It's just easier to add this variable than to update this url file
for every new version.

The reason to use Leap 15.4 image is because 15.5 is not yet released
and we use the stable version to be used as part of host tests.