Add test for apptainer package on TW
This package is maintained by HPC repos but it is not available yet for
HPC. However it is on factory and we want to follow the Factory first policy
and cover it before make it through on SLE.

The module covers basic operation of the tool. Apptainer is the new name of
singularity.

Signed-off-by: ybonatakis <ybonatakis@suse.com>