Add SELINUX=0 to tw upgrade and zdup tests (bsc#1230118)
except to tw2tw zdup tests.

We want to toggle SELINUX=1 as default in tw tests to be able
to test SELinux when it will be set as default in the installer.
However, when going from Leap to tw, AppArmor should be preserved
and tested, as we don't change the main LSM in these cases.

see:
https://github.com/os-autoinst/os-autoinst-distri-opensuse/pull/20125